Adicionado pacote para lidar com exceções de: campo não preenchido e recurso não encontrado.

Alguns métodos das classes ClienteController e ClienteService foram modificados para utilizarem das novas Exceptions.

obs: teste de buscar cliente vai falhar.

+++ src/main/java/com/unifacs/GQS_A3/Service/ClienteService.java
@@ -1,6 +1,8 @@
 package com.unifacs.GQS_A3.Service;
 
 import com.unifacs.GQS_A3.Repository.ClienteRepository;
+import com.unifacs.GQS_A3.exceptions.CampoNaoPreenchidoException;
+import com.unifacs.GQS_A3.exceptions.RecursoNaoEncontradoException;
 import com.unifacs.GQS_A3.model.Cliente;
 import org.springframework.stereotype.Service;
 
@@ -20,14 +22,21 @@ public class ClienteService {
         return clienteRepository.findAll();
     }
 
-    public Optional<Cliente> buscarPorId(Long id){
-        return clienteRepository.findById(id);
+    public Cliente buscarPorId(Long id){
+        return clienteRepository.findById(id).
+                orElseThrow(() -> new RecursoNaoEncontradoException("Cliente com id "+id+" não encontrado"));
     }
 
     public Cliente registrarCliente(Cliente cliente){
+        if(cliente.getNome() == null || cliente.getEmail() == null || cliente.getSenha() == null){
+            throw new CampoNaoPreenchidoException("nome, email e senha são obrigatórios");
+        }
         return clienteRepository.save(cliente);
     }
     public void removerCliente(Long id){
+        if(!clienteRepository.existsById(id)){
+            throw new RecursoNaoEncontradoException("Cliente com id "+id+" não encontrado");
+        }
         clienteRepository.deleteById(id);
     }
